# docs/swagger.yaml
basePath: /
definitions:
  domain.CreateTaskRequest:
    description: Данные для создания новой задачи
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  domain.CreateTaskResponse:
    description: UUID созданной задачи
    properties:
      id:
        type: string
    type: object
  domain.Task:
    description: Задача с UUID, заголовком, описанием, статусом и временными метками
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        $ref: '#/definitions/domain.TaskStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  domain.TaskListItem:
    description: Краткая информация о задаче для списка
    properties:
      id:
        type: string
      status:
        $ref: '#/definitions/domain.TaskStatus'
      title:
        type: string
    type: object
  domain.TaskStatus:
    enum:
    - new
    - in_progress
    - done
    type: string
    x-enum-varnames:
    - TaskStatusNew
    - TaskStatusInProgress
    - TaskStatusDone
  domain.UpdateTaskRequest:
    description: Данные для обновления задачи
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  domain.UpdateTaskResponse:
    description: Статус операции обновления
    properties:
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    REST API для управления задачами (CRUDL)
    Реализовано на чистом Go с использованием стандартной библиотеки
  title: Tasks API
  version: "1.0"
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Возвращает список задач с фильтрацией по статусу и пагинацией
      parameters:
      - description: Фильтр по статусу (new, in_progress, done)
        in: query
        name: status
        type: string
      - description: Лимит записей (по умолчанию 100, максимум 1000)
        in: query
        name: limit
        type: integer
      - description: Смещение для пагинации
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TaskListItem'
            type: array
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список задач
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Создаёт новую задачу с указанным заголовком и описанием
      parameters:
      - description: Данные задачи
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.CreateTaskResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать задачу
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет задачу по её UUID (идемпотентная операция)
      parameters:
      - description: UUID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Неверный UUID
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить задачу
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Возвращает задачу по её UUID
      parameters:
      - description: UUID задачи
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Task'
        "400":
          description: Неверный UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить задачу
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Обновляет данные задачи (заголовок, описание, статус)
      parameters:
      - description: UUID задачи
        in: path
        name: id
        required: true
        type: string
      - description: Обновлённые данные
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UpdateTaskResponse'
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Задача не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить задачу
      tags:
      - tasks
schemes:
- http
swagger: "2.0"